Changed Withdrawable to input a double amount instead of an int

+++ phase2/IWithdrawable.java
@@ -1,3 +1,3 @@
 public interface IWithdrawable {
-    double withdraw(int amount, double balance);
+    double withdraw(double amount, double balance);
 }
