+++ phase2/WithdrawCappedNonNegative.java
@@ -6,10 +6,10 @@ public class WithdrawCappedNonNegative implements IWithdrawable, Serializable {
         this.minBalance = minBalance;
     }
 
-    public double withdraw(int amount, double balance){
-        if (balance > 0 && balance + minBalance >= amount){
-            ATM_machine.withdrawBills(amount);
-            return balance - amount;
+    public double withdraw(double amount, double balance){
+        if (balance > 0 && balance + minBalance >= (int)amount){
+            ATM_machine.withdrawBills((int)amount);
+            return balance - (int)amount;
         }
         else {
             System.out.println("You don't have enough money remaining to withdraw that much! Balance: " + balance);
