+++ phase2/WithdrawUncapped.java
@@ -2,8 +2,8 @@ import java.io.Serializable;
 
 public class WithdrawUncapped implements IWithdrawable, Serializable {
 
-    public double withdraw(int amount, double balance){
-        ATM_machine.withdrawBills(amount);
-        return balance - amount;
+    public double withdraw(double amount, double balance){
+        ATM_machine.withdrawBills((int)amount);
+        return balance - (int)amount;
     }
 }
